+++ devapp-2.16/client/main.jsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { Meteor } from 'meteor/meteor'
 import { render } from 'react-dom'
 import { App } from '../imports/ui/App'
+import './local-collections'
 
 import '../imports/api/links'
 import '../imports/api/random'