+++ devapp-3.5/client/main.jsx
@@ -2,6 +2,7 @@ import { createRoot } from 'react-dom/client'
 import { Meteor } from 'meteor/meteor'
 import { App } from '/imports/ui/App'
 import '/imports/ui/styles.css'
+import './local-collections'
 
 Meteor.startup(() => {
   const container = document.getElementById('react-target')